+++ changelog.md
@@ -1,4 +1,12 @@
-v1.0.0
+v1.0.1
 
-first release
+Major improvement
+
+Introducing, we use seedtest.net's API.
+
+Download and upload speeds, accurate to your network.
+
+Automatically finds the nearest server.
+
+Fixed some addon crashing errors.